HACK: Limit SPI interrupts to 128
This prevents ICI traffic as IRQs above this are on remote nodes, and
the remote nodes in TM do not respond to GIC messages.  This fixes
the "Gone fishing" errors that are caused when the response never comes
and the code times out.

This hack is limited to Vulcan CPUs only and enabled via
CONFIG_BRCM_ERRATUM_SPI128.

DO NOT UPSTREAM